extracted store-related models to new separate file (renamed from store-finder)
