+++ projects/core/src/model/order.model.ts
@@ -1,11 +1,5 @@
 import { Price, Product } from './product.model';
-import {
-  GeoPoint,
-  OpeningSchedule,
-  PaginationModel,
-  SortModel,
-} from './misc.model';
-import { Image } from './image.model';
+import { PaginationModel, SortModel } from './misc.model';
 import { Address } from './address.model';
 import {
   DeliveryOrderEntryGroup,
@@ -14,28 +8,13 @@ import {
   PromotionResult,
   Voucher,
 } from './cart.model';
+import { PointOfService } from './store.model';
 
 export interface DeliveryMode {
   code?: string;
   deliveryCost?: Price;
   description?: string;
   name?: string;
-}
-
-export interface PointOfService {
-  address?: Address;
-  description?: string;
-  displayName?: string;
-  distanceKm?: number;
-  features?: { [propertyName: string]: string };
-  formattedDistance?: string;
-  geoPoint?: GeoPoint;
-  mapIcon?: Image;
-  name?: string;
-  openingHours?: OpeningSchedule;
-  storeContent?: string;
-  storeImages?: Image[];
-  url?: string;
 }
 
 export interface OrderEntry {
